WEBAPP: Use Quart instead of Flask[async]
Flask[async] was a little broken and hard to use with an async/await setup. Quart is a reimplementation of Flask using async/await.

This moves the load and remove threads to be async tasks on the same loop as the server instead.

One downside is that Quart's hot reloading doesn't seem to work correctly, and stops the process instead of restarting it.
So for now I set `use_reloading=False`.